Move {read/write}Mesh{From/To}File.{h,cpp}.
In order to make OGS more modular:
- The {read/write}Mesh{From/To}File.{h,cpp} are moved from FileIO to MeshLib/IO.
- The FileIO/Legacy folder was moved to MeshLib/IO/Legacy.
- Include paths are changed according the changes above.
