+++ Applications/Utils/FileConverter/GMSH2OGS.cpp
@@ -32,7 +32,7 @@
 
 // FileIO
 #include "FileIO/GMSHInterface.h"
-#include "FileIO/Legacy/MeshIO.h"
+#include "MeshLib/IO/Legacy/MeshIO.h"
 #include "FileIO/VtkIO/VtuInterface.h"
 
 // MeshLib
