+++ Applications/Utils/MeshEdit/ResetPropertiesInPolygonalRegion.cpp
@@ -22,8 +22,8 @@
 #include "Applications/ApplicationsLib/LogogSetup.h"
 
 // FileIO
-#include "FileIO/readMeshFromFile.h"
-#include "FileIO/writeMeshToFile.h"
+#include "MeshLib/IO/readMeshFromFile.h"
+#include "MeshLib/IO/writeMeshToFile.h"
 
 // GeoLib
 #include "GeoLib/GEOObjects.h"
