+++ Applications/Utils/ModelPreparation/CMakeLists.txt
@@ -1,7 +1,6 @@
 add_executable(ComputeNodeAreasFromSurfaceMesh ComputeNodeAreasFromSurfaceMesh.cpp)
 set_target_properties(ComputeNodeAreasFromSurfaceMesh PROPERTIES FOLDER Utilities)
 target_link_libraries(ComputeNodeAreasFromSurfaceMesh
-	FileIO
 	MeshLib
 	MathLib
 	${OGS_VTK_REQUIRED_LIBS}
