+++ MeshLib/CMakeLists.txt
@@ -10,9 +10,11 @@ GET_SOURCE_FILES(SOURCES_ELEMENTS Elements)
 GET_SOURCE_FILES(SOURCES_EDITING MeshEditing)
 GET_SOURCE_FILES(SOURCES_GENERATORS MeshGenerators)
 GET_SOURCE_FILES(SOURCES_SEARCH MeshSearch)
+GET_SOURCE_FILES(SOURCES_IO IO)
+GET_SOURCE_FILES(SOURCES_IO_LEGACY IO/Legacy)
 
 GET_SOURCE_FILES(SOURCES_QUALITY MeshQuality)
-set(SOURCES ${SOURCES_MESHLIB} ${SOURCES_ELEMENTS} ${SOURCES_EDITING} ${SOURCES_GENERATORS} ${SOURCES_QUALITY} ${SOURCES_SEARCH})
+set(SOURCES ${SOURCES_MESHLIB} ${SOURCES_ELEMENTS} ${SOURCES_EDITING} ${SOURCES_GENERATORS} ${SOURCES_QUALITY} ${SOURCES_SEARCH} ${SOURCES_IO} ${SOURCES_IO_LEGACY})
 
 # Create the library
 add_library(MeshLib STATIC ${SOURCES})
