+++ MeshLib/IO/Legacy/MeshIO.h
@@ -73,4 +73,3 @@ private:
 } /* namespace FileIO */
 
 #endif /* MESHIO_H_ */
-
